requirements.txt
python-dotenv==1.0.0
openai==1.3.0
anthropic==0.7.1
requests==2.31.0
